+++ src/lib/ssh_clients/ssh_static_keys_provider.py
@@ -20,6 +20,7 @@ class SSHStaticKeysProvider(SSHKeysProvider):
             return {
                 "private": self.users_keys[username].private_key.get_secret_value(),
                 "public": self.users_keys[username].public_cert,
+                "passphrase": self.users_keys[username].passphrase,
             }
         else:
             raise SSHCredentials(f"No SSH credentials found for user:{username}")